dom/imptests/html/html/dom/documents/dom-tree-accessors/test_document.title-06.html

<!doctype html>
<title>document.title and the empty string</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#document.title">
<meta name="assert" content="On setting document.title, a text node must be created.">
<link rel="stylesheet" href="/resources/testharness.css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  document.documentElement.firstChild.removeChild(
    document.documentElement.firstChild.firstChild);
  assert_equals(document.title, "");
  document.title = "";
  assert_equals(document.title, "");
  assert_true(document.documentElement.firstChild.lastChild instanceof
              HTMLTitleElement, "Need a title element.");
  var title = document.documentElement.firstChild.lastChild.firstChild;
  assert_true(title, "Need a node.");
  assert_true(title instanceof Text, "Should be a text node");
  assert_equals(title.value, "");
});
</script>
